benchmark: RC1 protocol, ledger, report template and stats that the template needs
Writes the two documents the RC1 prompt asks for before the campaign decides
anything, and the report path that renders them:

* docs/bench/rc1-130/PROTOCOL.md — the pre-registered rules as actually
  implemented: what each pass measures and what is excluded (competitors'
  Windows production path is NOT MEASURED, and no number here may be read as an
  end-to-end comparison), the paired/rotated design, medians-primary with tails
  published, the A/A floor and the tier vocabulary, the attribution-pair guard,
  the candidate acceptance rule, the gate list, the NOT AVAILABLE register, and
  a deviations log naming every instrument defect found this cycle with its
  commit.
* docs/bench/rc1-130/OPTIMIZATION_LEDGER.md — C-A (mechanism, the exhaustive
  walk oracle, the correctness argument, files), the specs and falsifiers for
  C-B/C-C/C-D/C-E, and the rejected list (R1-R5, C8) with reasons. The C-A
  measurement block is explicitly pending rc1-ca4 rather than filled with
  remembered numbers: the earlier rc1-ca*/rc1-baseline runs left no artifacts in
  the repo, so this campaign is the one the ledger will quote.
* benchmark/REPORT.rc1.narrative.md + make_report.py --narrative/--strict-stat —
  the RC1 report renders from a campaign's summary.json/tables.md/gates.txt only;
  strict mode turns an unresolvable {{stat:…}} into a build failure instead of a
  sentence containing 'n/a'.
* rc1_stats.py: candidate_verdict (the acceptance rule computed, not typed),
  latency_cells medians, per-engine correctness rates, the memory block, the
  diffab/robust/correctness totals, the manifest read for the toolchain line, and
  the docstring corrected to say 1x the A/A median for a regressing cell
  (2x is used only by the candidate rule, and both constants are now stated).

+++ benchmark/harness/rc1.hpp
@@ -445,8 +445,12 @@ inline int modeTimer(const Args& a, std::FILE* out) {
     j.addF("pair_max_ns", static_cast<double>(pairAg.mx), 0);
     j.addF("pair_mean_ns", pairAg.mean(), 2);
     j.end();
+    // pct() returns an integer count of nanoseconds: it must be cast before it
+    // meets %f, or the varargs read prints 0 (the JSON artifact was always right;
+    // a display-only defect still misleads the person watching a campaign run).
     std::printf("[timer] clock pair: res=%s min=%llu med=%.0f p99=%.0f ns\n", buf,
-                (unsigned long long)pairAg.mn, pairAg.pct(0.50), pairAg.pct(0.99));
+                (unsigned long long)pairAg.mn, static_cast<double>(pairAg.pct(0.50)),
+                static_cast<double>(pairAg.pct(0.99)));
     return 0;
 }
